Ketvirtoji versija

+++ mokinys.h
@@ -9,7 +9,7 @@ struct mokinys {
 	std::string vardas = "";
 	std::string pavarde = "";
 	std::vector<int> pazym;
-	int egz = 0; //egzamino pažymys
+	double egz = 0; //egzamino pažymys
 	double vidurkis = 0; //namų darbų pažymių vidurkis
 	double mediana = 0; //namų darbų pažymių mediana
